Authenticate Routes and Add Functionalities

+++ controllers/contactController.js
@@ -3,29 +3,34 @@ const Contact = require('../models/contactModel');
 
 // GET ALL CONTACTS
 const getContacts =asyncHandler(async (req,res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.user.id});
     res.status(200).json(contacts)
 });
 
 
 // POST METHOD CREATE CONTACT
 const createContact =asyncHandler(async (req,res) => {
-    console.log(req.body);
-    const {name, email,branch, password}= req.body;
+    const {name, email, phone} = req.body;
 
-    if(!name || !email || !branch|| !password){
+    console.log(name,email,phone);
+
+    if(!name || !email || !phone){
         res.status(400);
-        throw new Error('All fields are mandatory.')
+        throw new Error("All fields are mandatory");
     }
 
     const contact = await Contact.create({
         name,
         email,
-        branch,
-        password,
-        statusApproval : false
+        phone,
+        user_id: req.user.user.id
     })
-    res.status(201).json(contact)
+
+    console.log(contact);
+    res.send(contact)
+
+    // console.log(req.user);
+    // console.log(req.body);
 });
 
 // GET CONTACT WITH UNIQUE ID
@@ -48,6 +53,11 @@ const updateContact =asyncHandler( async (req,res) => {
         throw new Error("Contact not found")
     }
 
+    if(contact.user_id.toString() != req.user.user.id) {
+        res.status(403);
+        throw new Error("Unauthorized user")
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -63,6 +73,13 @@ const deleteContact =asyncHandler( async (req,res) => {
     if(!contact) {
         res.status(404);
         throw new Error("Contact not found")
+    }
+
+    console.log(contact.user_id);
+    console.log(req.user.user.id);
+    if(contact.user_id.toString() != req.user.user.id) {
+        res.status(403);
+        throw new Error("Unauthorized user")
     }
 
     await Contact.findByIdAndRemove(req.params.id);
